LPS-63097 Move portal-soap-extender to portal-remote-soap-extender
