+++ modules/portal-remote/portal-remote-soap-extender/bnd.bnd
@@ -4,12 +4,12 @@ Bundle-ClassPath:\
 	lib/cxf-rt-frontend-jaxws.jar,\
 	lib/cxf-rt-frontend-simple.jar,\
 	lib/cxf-rt-ws-policy.jar
-Bundle-Name: Liferay Portal SOAP Extender
-Bundle-SymbolicName: com.liferay.portal.soap.extender
+Bundle-Name: Liferay Portal Remote SOAP Extender
+Bundle-SymbolicName: com.liferay.portal.remote.soap.extender
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.portal.soap.extender,
-	com.liferay.portal.soap.extender.configuration
+	com.liferay.portal.remote.soap.extender,
+	com.liferay.portal.remote.soap.extender.configuration
 Import-Package:\
 	!com.sun.*,\
 	!org.springframework.*,\
