+++ modules/portal-remote/portal-remote-soap-extender/src/main/java/com/liferay/portal/remote/soap/extender/configuration/SoapExtenderConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.soap.extender.configuration;
+package com.liferay.portal.remote.soap.extender.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
 
@@ -26,7 +26,7 @@ import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClass
 )
 @Meta.OCD(
 	factory = true,
-	id = "com.liferay.portal.soap.extender.configuration.SoapExtenderConfiguration",
+	id = "com.liferay.portal.remote.soap.extender.configuration.SoapExtenderConfiguration",
 	localization = "content/Language",
 	name = "%soap.extender.configuration.name"
 )
