+++ modules/portal-remote/portal-remote-soap-extender/src/main/java/com/liferay/portal/remote/soap/extender/internal/SoapExtender.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.soap.extender.internal;
+package com.liferay.portal.remote.soap.extender.internal;
 
 import aQute.bnd.annotation.metatype.Configurable;
 
 import com.liferay.portal.remote.dependency.manager.tccl.TCCLDependencyManager;
-import com.liferay.portal.soap.extender.SoapDescriptorBuilder;
-import com.liferay.portal.soap.extender.configuration.SoapExtenderConfiguration;
+import com.liferay.portal.remote.soap.extender.SoapDescriptorBuilder;
+import com.liferay.portal.remote.soap.extender.configuration.SoapExtenderConfiguration;
 
 import java.util.Map;
 
@@ -41,7 +41,7 @@ import org.osgi.service.http.whiteboard.HttpWhiteboardConstants;
  * @author Carlos Sierra Andrés
  */
 @Component(
-	configurationPid = "com.liferay.portal.soap.extender.configuration.SoapExtenderConfiguration",
+	configurationPid = "com.liferay.portal.remote.soap.extender.configuration.SoapExtenderConfiguration",
 	configurationPolicy = ConfigurationPolicy.REQUIRE
 )
 public class SoapExtender {
