+++ modules/portal-remote/portal-remote-soap-extender/src/main/resources/content/Language_sk.properties
